shared: move base64, crc32, and misc utility functions
base64/crc32 already existed as separate near-identical or
functionally-equivalent copies in util/ and libnvme/src/nvme/ (libnvme
couldn't use the CLI's GPL copies, so it carried its own). Consolidate
on one copy each in shared/, verified to produce identical output to
what they replace:
 - base64: byte-for-byte identical code in both locations (Hannes
   Reinecke's original nvme-cli contribution, already properly
   relicensed to LGPL for the libnvme copy with his own
 - crc32: two independently-sourced implementations of the same
   algorithm (elfutils-derived vs. the 1986 Gary S. Brown/FreeBSD
   one), confirmed to produce identical checksums across the standard
   test vector and 1000+ fuzzed inputs. Keep the Gary S. Brown copy,
   already living happily in libnvme's LGPL codebase; its license
   text and attribution are preserved unchanged.

Also move a further batch of internal-only (non-public, no ABI
weight) helpers out of libnvme/src/nvme/util.c and util-fabrics.c.

Also deduplicate streq0/streqcase0/xstrdup, discovered to already be
duplicated between shared/string-util.h and libnvme/src/nvme/private.h
once the two collided in the same translation unit.
