+++ libnvme/src/meson.build
@@ -7,8 +7,6 @@
 #
 sources = [
     'nvme/accessors.c',
-    'nvme/base64.c',
-    'nvme/crc32.c',
     'nvme/ioctl.c',
     'nvme/lib.c',
     'nvme/log.c',
@@ -192,6 +190,7 @@ pkg.generate(libnvme,
 libnvme_dep = declare_dependency(
     link_with: libnvme,
     include_directories: '.',
+    dependencies: [shared_dep],
 )
 
 # test library with all symbols visible, to use for MI unit tests. Should
@@ -206,6 +205,7 @@ libnvme_test = library(
 libnvme_test_dep = declare_dependency(
     link_with: libnvme_test,
     include_directories: '.',
+    dependencies: [shared_dep],
 )
 
 mode = 'rw-r--r--'
