+++ shared/crc32.c
