+++ shared/meson.build
@@ -12,10 +12,22 @@
 
 sources = [
     'array-util.c',
+    'base64.c',
+    'crc32.c',
     'fs-util.c',
     'ini.c',
     'parse-util.c',
 ]
+
+# net-util.c wraps POSIX socket/interface APIs (ifaddrs.h, net/if.h) that
+# MinGW does not provide. Only NVMe-oF code calls into it, and fabrics
+# support is unavailable on Windows already (see want_fabrics), so it is
+# excluded rather than stubbed.
+if host_system != 'windows'
+    sources += [
+        'net-util.c',
+    ]
+endif
 
 libshared = static_library(
     'shared',
