+++ util/meson.build
@@ -2,8 +2,6 @@
 
 util_sources = [
     'util/argconfig.c',
-    'util/base64.c',
-    'util/crc32.c',
     'util/suffix.c',
     'util/types.c',
     'util/utils.c',
